+++ src/ble/gatt-service/battery_service_server.c
@@ -99,7 +99,7 @@ void battery_service_server_init(uint8_t value){
 	// get service handle range
 	uint16_t start_handle = 0;
 	uint16_t end_handle   = 0xfff;
-	int service_found = gatt_server_get_get_handle_range_for_service_with_uuid16(ORG_BLUETOOTH_SERVICE_BATTERY_SERVICE, &start_handle, &end_handle);
+	int service_found = gatt_server_get_handle_range_for_service_with_uuid16(ORG_BLUETOOTH_SERVICE_BATTERY_SERVICE, &start_handle, &end_handle);
 	btstack_assert(service_found != 0);
 	UNUSED(service_found);