+++ test/pts/gap_peripheral_test.c
@@ -1234,7 +1234,7 @@ int btstack_main(int argc, const char * argv[]){
     // report gatt service handle range for SDP tests
     uint16_t start_handle = 0;
     uint16_t end_handle   = 0xffff;
-    int service_found = gatt_server_get_get_handle_range_for_service_with_uuid16(ORG_BLUETOOTH_SERVICE_GENERIC_ATTRIBUTE, &start_handle, &end_handle);
+    int service_found = gatt_server_get_handle_range_for_service_with_uuid16(ORG_BLUETOOTH_SERVICE_GENERIC_ATTRIBUTE, &start_handle, &end_handle);
     btstack_assert(service_found != 0);
     printf("GATT Service range: 0x%04x - 0x%04x\n", start_handle, end_handle);